requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-community 
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-google-genai
streamlit
langchain_groq
duckduckgo-search
python-docx
PyPDF2
googlemaps
sentencepiece
sacremoses
indic-transliteration
transformers
google-generativeai